PolicyOps: move PenalizeAllNewlines
