+++ scalafmt-core/shared/src/main/scala/org/scalafmt/internal/Router.scala
@@ -61,6 +61,7 @@ class Router(formatOps: FormatOps) {
 
   import Constants._
   import LoggerOps._
+  import PolicyOps._
   import TokenOps._
   import TreeOps._
   import formatOps._
